Update App Architecture

+++ src/com/hideaway/RenderEngine/graphics/coordinates/matricies/Matrix.java
@@ -4,8 +4,8 @@ package com.hideaway.RenderEngine.graphics.coordinates.matricies;
  * which is a constant projection matrix that allows us to transform vertices in regards to our perspective.
  * */
 public class Matrix {
-    public float[][] mat4x4 = new float[4][4];
-    public Matrix(float[][] floats){
+    public double[][] mat4x4 = new double[4][4];
+    public Matrix(double[][] floats){
         //Must take a maximum of 16 floating point numbers, because this is a 4x4 matrix.
         for (int i = 0; i <=3; i++){
             for (int l = 0; l <=3; l++){
