+++ src/com/hideaway/RenderEngine/graphics/coordinates/matricies/RotationMatrix.java
@@ -1,7 +1,7 @@
 package com.hideaway.RenderEngine.graphics.coordinates.matricies;
 
 public class RotationMatrix extends Matrix {
-    public RotationMatrix(float[][] rotation) {
+    public RotationMatrix(double[][] rotation) {
         super(rotation);
     }
 }
