+++ src/com/hideaway/RenderEngine/graphics/coordinates/matricies/RotationMatrixX.java
@@ -1,12 +1,12 @@
 package com.hideaway.RenderEngine.graphics.coordinates.matricies;
 
 public class RotationMatrixX extends RotationMatrix {
-    public static float fTheta = 0;
-    public static float cosf = (float) Math.cos(fTheta);
-    public static float sinf = (float) Math.sin(fTheta);
+    public static double fTheta = 0.0d;
+    public static double cosf = (double) Math.cos(fTheta);
+    public static double sinf = (double) Math.sin(fTheta);
 
     public RotationMatrixX() {
-        super(new float[][] {
+        super(new double[][] {
                 {1,0,0,0},
                 {0,cosf,-sinf,0},
                 {0,sinf,cosf,0},
