+++ src/com/hideaway/RenderEngine/graphics/coordinates/matricies/RotationMatrixY.java
@@ -1,12 +1,12 @@
 package com.hideaway.RenderEngine.graphics.coordinates.matricies;
 
 public class RotationMatrixY extends RotationMatrix{
-    public static float fTheta = 0.f;
-    public static float cosf = (float) Math.cos(fTheta);
-    public static float sinf = (float) Math.sin(fTheta);
+    public static double fTheta = 0.0d;
+    public static double cosf = Math.cos(fTheta);
+    public static double sinf = Math.sin(fTheta);
 
     public RotationMatrixY() {
-        super(new float[][] {
+        super(new double[][] {
                 {cosf,0,sinf,0},
                 {0,1,0,0},
                 {-sinf,0,cosf,0},
