+++ src/com/hideaway/RenderEngine/graphics/coordinates/matricies/RotationMatrixZ.java
@@ -1,12 +1,12 @@
 package com.hideaway.RenderEngine.graphics.coordinates.matricies;
 
 public class RotationMatrixZ extends RotationMatrix {
-    public static float fTheta = 0.f;
-    public static float cosf = (float) Math.cos(fTheta);
-    public static float sinf = (float) Math.sin(fTheta);
+    public static double fTheta = 0.0d;
+    public static double cosf = (double) Math.cos(fTheta);
+    public static double sinf = (double) Math.sin(fTheta);
 
     public RotationMatrixZ() {
-        super(new float[][] {
+        super(new double[][] {
                 {cosf,-sinf,0,0},
                 {sinf,cosf,0,0},
                 {0,0,1,0},
